refactor: divide template content component
